# ENH: Use openmp on x86-simd-sort to speed up np.sort and np.argsort