@r-devulap 2025-04-02
@rgommers pointed out this discussion in scipy that details the complications of using openmp https://github.com/scipy/scipy/issues/10239#issuecomment-795030817