@rgommers 2025-04-02
I suspect that if we start using OpenMP code, we should disable it in wheels and only let distro packagers enable it.