@r-devulap 2025-04-09
> I suspect that if we start using OpenMP code, we should disable it in wheels and only let distro packagers enable it.

@rgommers How does the interaction with PyTorch or Sciki-learn work? Don't they vendor their own version of libgomp which can potentially conflict with what the distro provides?