@rgommers 2025-04-09
> @rgommers How does the interaction with PyTorch or Sciki-learn work? Don't they vendor their own version of libgomp which can potentially conflict with what the distro provides?

That is already the case - the answer for that one is: (a) please don't mix distro packages with wheels, and (b) the `libomp`/`libiomp` (not `libgomp`, which interacts badly with `multiprocessing`) inside wheels will have its symbols mangled with `auditwheel`, so there's no clash.

If you have PyTorch and scikit-learn both installed in the same environment and then imported, that usually works just fine, but it has given rise to a host of hard to debug issues in the past. Also `numpy` from conda defaults, which pulls in OpenMP via MKL. In wheels you just get two OpenMP runtimes loaded that are isolated from each other (bad for performance but should be robust again symbol conflicts), in conda envs you _should_ get only one but can get two if dependency trees don't work out well. E.g., PyTorch relies on MKL which pulls in Intel OpenMP (`libiomp`); scikit-learn uses LLVM OpenMP (`libomp`). There has to be code that handles this correctly, to avoid conflicts. IIRC `mkl-service` does this, so it uses the other OpenMP runtime if that's already loaded in memory.

You get the gist - this is a little painful. There's only one way to really do OpenMP right - build everything in a coherent fashion against the same OpenMP library. Distros usually get this right. With wheels you can't. And it gets worse when users do `pip install .` or `pip install pkgs-that-uses-openmp` from source - because then you don't get `auditwheel` symbol mangling, and things may go more wrong.

* * * 

This PR looks nice and simple though, so if the performance gains are really large, maybe making it opt-in rather than use-if-detected could work.