@r-devulap 2025-04-14
updated patch: 

1. Ported openMP support for `np.argsort` from https://github.com/intel/x86-simd-sort/pull/195 which speeds up sorting arrays > 10000 by up-to 3.5x on both AVX-512 and AVX2. Updated the title and added benchmark numbers in the first comment.
2. Added a new meson option `-enable-openmp` which is false by default. Meson only builds with openmp when both `disable-threading == false` and `enable-openmp == true`
3. Added CI coverage to test the openMP code paths.