@rgommers 2025-04-16
This will also need a release note in `doc/release/upcoming_changes/`