@r-devulap 2025-04-17
> scikit-learn does run into issues with running OpenMP's threadpool together with OpenBLAS: https://github.com/scikit-learn/scikit-learn/issues/28883

@thomasjpfan Thanks for the reference. Correct me if I am wrong but from what I understand, it looks like the performance problem occurs when calling two functions back to back in a loop: where one uses pthreads and other uses openMP to manage their respective threads. This causes resource contention when both functions try to use available CPU cores with no visibility into what the other one is doing. 

Beyond standardizing the thread management library across the entire Python ecosystem, are there alternative approaches to solve this?