@thomasjpfan 2025-04-17
> This causes resource contention when both functions try to use available CPU cores with no visibility into what the other one is doing.

Yes, this is the underlying issue.

> Beyond standardizing the thread management library across the entire Python ecosystem, are there alternative approaches to solve this?

Standardizing the thread management layer is the only long term solution I see. The hard part is figuring out a way to standardize that works for most projects. Top of mind projects and how they multi-thread:

- SciPy's fft uses C++ threads & OpenBLAS
- NumPy uses OpenBLAS
- PyTorch uses MKL
- Polars uses Rust+rayon
- Scikit-learn uses OpenMP and Python multi-threading
- Python multi-threading uses pthreads, which will become more common with free-threading

---

Some workarounds for threadpool specific issues:

- Contention: Set `OPENBLAS_THREAD_TIMEOUT`: 
    - https://github.com/OpenMathLib/OpenBLAS/issues/3187 
    - https://github.com/scipy/scipy/issues/15129
- Incompatible between Intel OpenMP and LLVM OpenMP: Set `MKL_THREADING_LAYER `
    - https://github.com/joblib/threadpoolctl/blob/master/multiple_openmp.md