@gitboy16 2025-05-20
Hi, thank you for the PR. Would it be possible to have wheels/packages with opennmp enabled available somewhere so that multithreaded sort can be used? Thank you!