@rgommers 2025-05-20
That would a lot of work including vendoring `libomp` into wheels, we won't be putting anything like that up on PyPI. If you or anyone else is willing to do this in a fork and host them outside of PyPI, that is of course possible - there's no need for us as maintainers to do that.